+++ test/org/jetbrains/plugins/ideavim/VimTestCase.java
@@ -1,6 +1,7 @@
 package org.jetbrains.plugins.ideavim;
 
 import com.intellij.openapi.application.PathManager;
+import com.intellij.openapi.editor.Caret;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.fileTypes.PlainTextFileType;
 import com.intellij.openapi.project.Project;
@@ -22,6 +23,7 @@ import com.maddyhome.idea.vim.helper.StringHelper;
 import com.maddyhome.idea.vim.option.Options;
 import com.maddyhome.idea.vim.ui.ExEntryPanel;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
 import java.util.ArrayList;
@@ -113,9 +115,12 @@ public abstract class VimTestCase extends UsefulTestCase {
     return keys;
   }
 
-  public void assertOffset(int expectedOffset) {
-    final int offset = myFixture.getEditor().getCaretModel().getOffset();
-    assertEquals(expectedOffset, offset);
+  public void assertOffset(int... expectedOffsets) {
+    final List<Caret> carets = myFixture.getEditor().getCaretModel().getAllCarets();
+    assertEquals("Wrong amount of carets", expectedOffsets.length, carets.size());
+    for (int i = 0; i < expectedOffsets.length; i++) {
+      assertEquals(expectedOffsets[i], carets.get(i).getOffset());
+    }
   }
 
   public void assertMode(@NotNull CommandState.Mode expectedMode) {
@@ -123,7 +128,7 @@ public abstract class VimTestCase extends UsefulTestCase {
     assertEquals(expectedMode, mode);
   }
 
-  public void assertSelection(@NotNull String expected) {
+  public void assertSelection(@Nullable String expected) {
     final String selected = myFixture.getEditor().getSelectionModel().getSelectedText();
     assertEquals(expected, selected);
   }